requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
gunicorn==20.1.0
jmespath==0.10.0
pdfkit==0.6.0
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
